+++ nengo/conftest.py
@@ -269,6 +269,57 @@ def seed(request):
     return function_seed(request.function, mod=TestConfig.test_seed)
 
 
+@pytest.fixture
+def allclose(request):
+    """A functional replacement for np.allclose.
+
+    This allows backends to specify different tolerances by adding
+
+    .. code-block:: ini
+
+        nengo_test_tolerances =
+            path/to/test/file:testname atol=x rtol=y
+            path/to/other/file:testname2 atol=z  # this is a comment
+            ...
+
+    to their pytest config file.
+
+    If a test contains multiple ``allclose`` checks then multiple entries
+    with the same testname can be added to the config, and will be applied
+    in the given order.
+    """
+
+    call_count = [0]  # need to access the closure this way for 2.7 compat
+    override = []
+    for line in request.config.getini("nengo_test_tolerances"):
+        split_line = line.split()
+        testname = "*" + split_line[0]
+        nodename = get_item_name(request.node)
+        if fnmatch(nodename, testname):
+            kwargs = {}
+            for entry in split_line[1:]:
+                if entry.startswith("#"):
+                    break
+                k, v = entry.split("=")
+                kwargs[k] = float(v)
+            override.append(kwargs)
+
+    def _allclose(a, b, **kwargs):
+        # set tolerances
+        if override:
+            kwargs.update(override[min(call_count[0], len(override) - 1)])
+            call_count[0] += 1
+
+        # record rmse
+        rmse = npext.rmse(a, b)
+        if not np.any(np.isnan(rmse)):
+            request.node.user_properties.append(("rmse", rmse))
+
+        return np.allclose(a, b, **kwargs)
+
+    return _allclose
+
+
 def pytest_generate_tests(metafunc):
     marks = [
         getattr(pytest.mark, m.name)(*m.args, **m.kwargs)